+++ sdk/eventhub/event-hubs/samples/javascript/websockets.js
@@ -1,7 +1,7 @@
-/*
-  Copyright (c) Microsoft Corporation. All rights reserved.
-  Licensed under the MIT Licence.
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT Licence.
 
+/*
   This sample demonstrates how to use WebSockets enable Event Hubs to work over an HTTP proxy and
   in environments where the standard AMQP port 5671 is blocked. For the latter case, ignore proxy
   related configurations in this sample.
@@ -14,11 +14,14 @@
   https://github.com/Azure/azure-sdk-for-js/tree/%40azure/event-hubs_2.1.0/sdk/eventhub/event-hubs/samples instead.
 */
 
-import { runSample } from './sampleHelpers';
-import { EventHubConsumerClient } from "@azure/event-hubs";
-import WebSocket from "ws";
+const WebSocket = require("ws");
 const url = require("url");
 const httpsProxyAgent = require("https-proxy-agent");
+
+const { EventHubConsumerClient } = require("@azure/event-hubs");
+
+// Load the .env file if it exists
+require("dotenv").config();
 
 // Define connection string and related Event Hubs entity name here
 const connectionString = process.env["EVENTHUB_CONNECTION_STRING"] || "";
@@ -32,21 +35,26 @@ const urlParts = url.parse("http://localhost:3128");
 urlParts.auth = "username:password"; // Skip this if proxy server does not need authentication.
 const proxyAgent = new httpsProxyAgent(urlParts);
 
-export async function main(): Promise<void> {
+async function main() {
   console.log(`Running websockets sample`);
 
   const client = new EventHubConsumerClient(consumerGroup, connectionString, eventHubName, {
     webSocketOptions: {
       webSocket: WebSocket,
       webSocketConstructorOptions: { agent: proxyAgent }
-    }    
+    }
   });
-  /*
-   Refer to other samples, and place your code here to send/receive events
-  */
-  await client.close();
 
+  /*
+     Refer to other samples, and place your code here to send/receive events
+    */
+
+  await client.close();
   console.log(`Exiting websockets sample`);
 }
 
-runSample(main);
+module.exports = { main };
+
+main().catch((error) => {
+  console.error("Error running sample:", error);
+});