+++ sdk/storage/storage-queue/src/QueueClient.ts
@@ -6,7 +6,8 @@ import {
   TokenCredential,
   isTokenCredential,
   isNode,
-  getDefaultProxySettings
+  getDefaultProxySettings,
+  URLBuilder
 } from "@azure/core-http";
 import { CanonicalCode } from "@opentelemetry/types";
 import {
@@ -38,14 +39,13 @@ import {
   appendToURLPath,
   extractConnectionStringParts,
   truncatedISO8061Date,
-  getValueInConnString,
   getStorageClientContext
 } from "./utils/utils.common";
 import { StorageSharedKeyCredential } from "./credentials/StorageSharedKeyCredential";
 import { AnonymousCredential } from "./credentials/AnonymousCredential";
 import { createSpan } from "./utils/tracing";
-import { DevelopmentConnectionString } from "./utils/constants";
 import { Metadata } from "./models";
+import { getCachedDefaultHttpClient } from "./utils/cache";
 
 /**
  * Options to configure {@link QueueClient.create} operation
@@ -497,7 +497,12 @@ export class QueueClient extends StorageClient {
       | string,
     options?: StoragePipelineOptions
   ) {
-    options = options || {};
+    // when options.httpClient is not specified, passing in a DefaultHttpClient instance to
+    // avoid each client creating its own http client.
+    const newOptions: StoragePipelineOptions = {
+      httpClient: getCachedDefaultHttpClient(),
+      ...options
+    };
     let pipeline: Pipeline;
     let url: string;
     if (credentialOrPipelineOrQueueName instanceof Pipeline) {
@@ -511,7 +516,7 @@ export class QueueClient extends StorageClient {
     ) {
       // (url: string, credential?: StorageSharedKeyCredential | AnonymousCredential | TokenCredential, options?: StoragePipelineOptions)
       url = urlOrConnectionString;
-      pipeline = newPipeline(credentialOrPipelineOrQueueName, options);
+      pipeline = newPipeline(credentialOrPipelineOrQueueName, newOptions);
     } else if (
       !credentialOrPipelineOrQueueName &&
       typeof credentialOrPipelineOrQueueName !== "string"
@@ -519,7 +524,7 @@ export class QueueClient extends StorageClient {
       // (url: string, credential?: StorageSharedKeyCredential | AnonymousCredential | TokenCredential, options?: StoragePipelineOptions)
       // The second paramter is undefined. Use anonymous credential.
       url = urlOrConnectionString;
-      pipeline = newPipeline(new AnonymousCredential(), options);
+      pipeline = newPipeline(new AnonymousCredential(), newOptions);
     } else if (
       credentialOrPipelineOrQueueName &&
       typeof credentialOrPipelineOrQueueName === "string"
@@ -534,15 +539,15 @@ export class QueueClient extends StorageClient {
             extractedCreds.accountKey
           );
           url = appendToURLPath(extractedCreds.url, queueName);
-          options.proxyOptions = getDefaultProxySettings(extractedCreds.proxyUri);
-          pipeline = newPipeline(sharedKeyCredential, options);
+          newOptions.proxyOptions = getDefaultProxySettings(extractedCreds.proxyUri);
+          pipeline = newPipeline(sharedKeyCredential, newOptions);
         } else {
           throw new Error("Account connection string is only supported in Node.js environment");
         }
       } else if (extractedCreds.kind === "SASConnString") {
         const queueName = credentialOrPipelineOrQueueName;
         url = appendToURLPath(extractedCreds.url, queueName) + "?" + extractedCreds.accountSas;
-        pipeline = newPipeline(new AnonymousCredential(), options);
+        pipeline = newPipeline(new AnonymousCredential(), newOptions);
       } else {
         throw new Error(
           "Connection string must be either an Account connection string or a SAS connection string"
@@ -1103,22 +1108,20 @@ export class QueueClient extends StorageClient {
       //  URL may look like the following
       // "https://myaccount.queue.core.windows.net/myqueue?sasString".
       // "https://myaccount.queue.core.windows.net/myqueue".
-      // or an emulator URL that starts with the endpoint `http://127.0.0.1:10001/devstoreaccount1`
+      // IPv4/IPv6 address hosts, Endpoints - `http://127.0.0.1:10001/devstoreaccount1/myqueue`
+      // http://localhost:10001/devstoreaccount1/queuename
 
-      let urlWithoutSAS = this.url.split("?")[0]; // removing the sas part of url if present
-      urlWithoutSAS = urlWithoutSAS.endsWith("/") ? urlWithoutSAS.slice(0, -1) : urlWithoutSAS; // Slicing off '/' at the end if exists
+      const parsedUrl = URLBuilder.parse(this.url);
 
-      // http://127.0.0.1:10001/devstoreaccount1
-      const emulatorQueueEndpoint = getValueInConnString(
-        DevelopmentConnectionString,
-        "QueueEndpoint"
-      );
-
-      if (this.url.startsWith(emulatorQueueEndpoint)) {
-        // Emulator URL starts with `http://127.0.0.1:10001/devstoreaccount1`
-        queueName = urlWithoutSAS.match(emulatorQueueEndpoint + "/([^/]*)")![1];
+      if (parsedUrl.getHost()!.split(".")[1] === "queue") {
+        // "https://myaccount.queue.core.windows.net/queuename".
+        // .getPath() -> /queuename
+        queueName = parsedUrl.getPath()!.split("/")[1];
       } else {
-        queueName = urlWithoutSAS.match("([^/]*)://([^/]*)/([^/]*)")![3];
+        // IPv4/IPv6 address hosts... Example - http://192.0.0.10:10001/devstoreaccount1/queuename
+        // Single word domain without a [dot] in the endpoint... Example - http://localhost:10001/devstoreaccount1/queuename
+        // .getPath() -> /devstoreaccount1/queuename
+        queueName = parsedUrl.getPath()!.split("/")[2];
       }
 
       if (!queueName) {