+++ sdk/storage/storage-file-share/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release History
 
+## 12.0.1 (Unreleased)
+
+- Share a single httpClient among clients. [PR #6657](https://github.com/Azure/azure-sdk-for-js/pull/6657)
+  Currently clients share their parent client's pipeline. But when the pipeline doesn't have a http client specified, a `DefaultHttpClient` will be created for each of the clients. This leads to memory leaks and exhaustion of TCP connections when many clients are created with `keepAlive` option enabled because each client creates a long-live http connection.  This change creates a default http client instance to share among all clients if none is specified in the pipeline options.
+
 ## 12.0.0 (2019-12-04)
 
 - [Breaking] The default browser bundle has been removed from the npm package. Bundling your application with a bundler such as Webpack is the recommended approach to building a browser bundle. For details on how to do this, please refer to our [bundling documentation](https://aka.ms/AzureSDKBundling).